Add project for the AD9144-FMC-EBZ + ZC706
Add a project for the AD9144-FMC-EBZ dual DAC board connected to the ZC706
carrier board.

This project assumes that the transceiver reference clock and SYSREF are
provided via the clock distribution chip that is found on the
AD9144-FMC-EBZ board. The reference clock rate must be the lane rate
divided by 40. SYSREF should be the reference clock divided by 64 (or a
larger power-of-two).

Due to system constraints the reference clock must be in the range of 75 to
165 MHz, which means the lane rate must be between 3 and 6.6 Gbps

This design can support the following FMC boards which are all pin
compatible:
  * AD9135-FMC-EBZ
  * AD9136-FMC-EBZ
  * AD9144-FMC-EBZ
  * AD9152-FMC-EBZ
  * AD9154-FMC-EBZ

Note that the AD9152-FMC-EBZ only uses the first 4 lanes, whereas all other
boards use 8 lanes.

The project is called dac_fmc_ebz as the intention is to support all DAC
FMC evaluation boards with this project since they are sufficiently similar
to be supported by the same design.

This project will successively extended to add support for more boards.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>