axi_adc_trigger: Manage trigger path
The axi_logic analyzer handles 3 output triggers:
- trigger_o - external trigger (no delay)
- trigger_out - signals on dmac/fifo_wr_sync the start of a new transfer.
A variable fifo depth is present in the data path, which delays the data
arriving at the DMA with 3 clock cycles. By coincidence, the external trigger
is synchronized and detected on 3 clock cycles. To get a maximum optimization
the trigger_out will be delayed with 3 clock cycles for internal triggers and
directly forwarded in the case of an external trigger.
- trigger_out_express (cascade trigger for logic_analyzer - m2k example)

Because the trigger_out_express must have a small delay, to get an exact
cascade trigger mechanism, a 1 delay clock cycle must be added on the data and
trigger paths, to avoid creating a closed combinatorial loop, leading to race
conditions.